Adjust priority queue max level and per-node capacity
The new max level of 8 means that the queue should now be suitable
only up to 2^8 elements, which is likely enough for most realistic
workloads. Though, empirical testing seems to indicate that a max
level of 6 to 8 performs better and is more consistent than the
previous maximum of 12. If needed, it can still be raised back.

The new per-node capacity is now 22 bit, down from 31. This makes
the node bitset to fit in 31 bit total, down from 44, while still
allowing some 4 million items per node.